verification token validation codes modified

+++ app/Http/Controllers/Api/VerifyEmailController.php
@@ -18,14 +18,38 @@ class VerifyEmailController extends Controller
             'token' => 'required|string',
         ]);
 
+        //count the number of tokens for this email
+        $countOtp = Otp::where('email', $request->email)
+        ->where('token', $request->token)
+        ->count();
+
+        if ($countOtp == 0) {
+            //status 0 means email is not verified
+            $status = 0;
+            return response()->json([
+                'message' => 'Invalid token',
+                'status' => $status
+
+            ], 400);
+        }
+
         //check if the email has this token
-        $otp = Otp::where('email', $request->email)->where('token', $request->token)->first();
+        $otp = Otp::where('email', $request->email)
+        ->where('token', $request->token)
+        ->first();
         
         //check if token is lasted more than 24 hours
         if ($otp->created_at->diffInHours(now()) > 24) {
             //delete the token
             $otp->delete();
-            return response()->json(['message' => 'Token is expired'], 400);
+
+            //status 0 means email is not verified
+            $status = 0;
+            return response()->json([
+                'message' => 'Token is expired',
+                'status' => $status
+
+            ], 400);
         }
 
         //if the email has this token
@@ -36,7 +60,15 @@ class VerifyEmailController extends Controller
             $user->save();
             //delete the token
             $otp->delete();
-            return response()->json(['message' => 'Email verified successfully'], 200);
+            
+            //status 1 means email is verified
+            $status = 1;
+            return response()->json([
+                'message' => 'Email verified successfully',
+                'status' => $status
+            ], 200);
         }
+
+       
     }
 }
